refactor: activity, view 패키지 구분
