+++ app/src/main/AndroidManifest.xml
@@ -12,13 +12,13 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".ReservationResultActivity"
+            android:name=".activity.ReservationResultActivity"
             android:exported="false" />
         <activity
-            android:name=".MovieReservationActivity"
+            android:name=".activity.MovieReservationActivity"
             android:exported="true" />
         <activity
-            android:name=".MainActivity"
+            android:name=".activity.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
