+++ app/src/main/java/woowacourse/movie/view/DateSpinner.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.view
 
 import android.content.Context
 import android.os.Bundle
