+++ app/src/main/java/woowacourse/movie/view/SaveState.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.view
 
 import android.os.Bundle
