+++ app/src/main/java/woowacourse/movie/view/SaveStateCounter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.view
 
 import android.os.Bundle
