+++ app/src/main/java/woowacourse/movie/view/SaveStateSpinner.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.view
 
 import android.R
 import android.content.Context
