+++ app/src/main/res/layout/activity_movie_reservation.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MovieReservationActivity">
+    tools:context=".activity.MovieReservationActivity">
 
     <ScrollView
         android:layout_width="match_parent"
