+++ app/src/main/res/layout/activity_reservation_result.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ReservationResultActivity">
+    tools:context=".activity.ReservationResultActivity">
 
     <TextView
         android:id="@+id/movie_reservation_result_information"
